# Modfiy hack host header parsing to use 0 allocs