[]byte->string and string->[]byte conversions tend to create lots of
garbage and are (relatively) slow compared to just sticking with bytes.
Also generic byte ops like `bytes.Split` are really expensive (although
less expensive than a regexp).

Since we already know the exact format we're expecting to get, let's the
parsing manually instead of relying on the bytes/strings package.

This makes the overhead of the header parsing nearly nil.

```
benchmark               old ns/op     new ns/op     delta
BenchmarkWithHack-8     67039         31338         -53.25%
BenchmarkNoHack-8       29701         29453         -0.83%

benchmark               old MB/s     new MB/s     speedup
BenchmarkWithHack-8     24.61        52.65        2.14x
BenchmarkNoHack-8       55.55        56.02        1.01x

benchmark               old allocs     new allocs     delta
BenchmarkWithHack-8     239            0              -100.00%
BenchmarkNoHack-8       0              0              +0.00%

benchmark               old bytes     new bytes     delta
BenchmarkWithHack-8     16309         0             -100.00%
BenchmarkNoHack-8       0             0             +0.00%
```

Signed-off-by: Brian Goff <cpuguy83@gmail.com>